+++ src/main/java/ca/corefacility/bioinformatics/irida/model/workflow/analysis/AnalysisOutputFile.java
@@ -1,5 +1,7 @@
 package ca.corefacility.bioinformatics.irida.model.workflow.analysis;
 
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Date;
 import java.util.Objects;
@@ -174,5 +176,16 @@ public class AnalysisOutputFile extends IridaResourceSupport implements IridaThi
 		}
 
 		return false;
+	}
+
+	/**
+	 * Read the bytes for an image output file
+	 *
+	 * @return the bytes for the file
+	 * @throws IOException if the file couldn't be read
+	 */
+	public byte[] getBytesForFile() throws IOException  {
+		byte[] bytes = Files.readAllBytes(getFile());
+		return bytes;
 	}
 }