+++ src/main/webapp/resources/js/components/Buttons/index.js
@@ -1,1 +1,3 @@
 export { ConfirmButton } from "./ConfirmButton";
+export { IconButton } from "./IconButton";
+export { RemoveTableItemButton } from "./RemoveTableItemButton";