+++ CHANGELOG.md
@@ -21,8 +21,25 @@ Changes
 * [UI]: Updated Administrator Users page to use Ant Design.
 * [UI]: Removed DataTables from overrepresented sequence file and run files page.
 * [REST]: Added a REST API endpoint to read assemblies for samples.
+* [UI]: Added selection for linker script to get assemblies in project/samples page.
 * [Developer]: `bower` completely removed.
 * [UI/Developer]: Updated analysis results page with Ant Design and React.
+* [UI]: Assemblies can now be uploaded through the sample files page.
+* [UI/Developer]: Updated to `Prettier` 2.0.2
+* [UI/Developer]: Updated to `eslint-plugin-react-hooks` to 3.0.0.
+* [UI/Developer]: Removed `angular-ui-bootstrap` as a global dependency (still required on several pages).
+* [UI/Developer]: New login page with developer ability to override template.
+* [UI/Developer]: Added excel file viewer to analysis outputs view.
+* [UI]: Added selection to export assemblies to Galaxy.
+* [UI]: Updated Project > Members page to remove `DataDables` and use Ant Design.
+* [UI/Developer]: Added image file viewer to analysis outputs view.
+* [UI]: Moved project > details into the project > settings panel.
+* [UI]: Updated Project navigation to use Ant Design menu.
+* [UI/Developer]: Updated to JQuery v3.5.1 to fix security vulnerability.
+* [UI]: Fixed bug preventing the upload of `*.fastqc.gz` on Mac computers.
+* [UI]: Update Groups page to remove `DataTables` and use Ant Design.
+* [UI]: Update Groups details page to use Ant Design and remove `DataTables`.
+* [UI]: Fixed bug preventing display of execution tools for first file if a provenance for another file had been viewed.
 * [UI/Developer]: Created a Thymeleaf `WebpackerDialect` for dynamically adding front end assets to html pages. See [Webpack Documentation](https://irida.corefacility.ca/documentation/developer/interface/webpack/) for more information.
 
 19.09 to 20.01